# PHP_ICANH_Project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Vehículos - ICANH",
        "description": "API RESTful para la gestión de vehículos, marcas, personas y sus relaciones",
        "contact": {
            "name": "Jhoan Sebastian Wilches Jimenez",
            "email": "sebastianwilches2@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/marcas-vehiculo": {
            "get": {
                "tags": [
                    "Marcas de Vehículo"
                ],
                "summary": "Obtener todas las marcas de vehículo",
                "operationId": "a8da97f43bd8fa170f113ece7becaeba",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "Número de registros a saltar (paginación)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número máximo de registros a devolver",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de marcas de vehículo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre_marca": {
                                                "type": "string",
                                                "example": "Toyota"
                                            },
                                            "pais": {
                                                "type": "string",
                                                "example": "Japón"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Marcas de Vehículo"
                ],
                "summary": "Crear una nueva marca de vehículo",
                "operationId": "bd3a9ff8fa601e9802eafc759ed1da80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_marca",
                                    "pais"
                                ],
                                "properties": {
                                    "nombre_marca": {
                                        "type": "string",
                                        "minLength": 1,
                                        "example": "Toyota"
                                    },
                                    "pais": {
                                        "type": "string",
                                        "minLength": 1,
                                        "example": "Japón"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Marca creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre_marca": {
                                            "type": "string",
                                            "example": "Toyota"
                                        },
                                        "pais": {
                                            "type": "string",
                                            "example": "Japón"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/marcas-vehiculo/{marca_vehiculo}": {
            "get": {
                "tags": [
                    "Marcas de Vehículo"
                ],
                "summary": "Obtener una marca de vehículo por ID",
                "operationId": "91c1b363bc1eb56184fd651074b6cdee",
                "parameters": [
                    {
                        "name": "marca_vehiculo",
                        "in": "path",
                        "description": "ID de la marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre_marca": {
                                            "type": "string",
                                            "example": "Toyota"
                                        },
                                        "pais": {
                                            "type": "string",
                                            "example": "Japón"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Marca no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Marcas de Vehículo"
                ],
                "summary": "Actualizar una marca de vehículo",
                "operationId": "282a502aa72133139489b8a4d34ec000",
                "parameters": [
                    {
                        "name": "marca_vehiculo",
                        "in": "path",
                        "description": "ID de la marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_marca": {
                                        "type": "string",
                                        "minLength": 1,
                                        "example": "Toyota Updated"
                                    },
                                    "pais": {
                                        "type": "string",
                                        "minLength": 1,
                                        "example": "Japón Updated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marca actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre_marca": {
                                            "type": "string",
                                            "example": "Toyota"
                                        },
                                        "pais": {
                                            "type": "string",
                                            "example": "Japón"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Marca no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Marcas de Vehículo"
                ],
                "summary": "Eliminar una marca de vehículo",
                "operationId": "4ea77ad65c67a0375433b3730f040984",
                "parameters": [
                    {
                        "name": "marca_vehiculo",
                        "in": "path",
                        "description": "ID de la marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Marca eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No se puede eliminar por tener vehículos asociados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se puede eliminar la marca porque tiene vehículos asociados"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Marca no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}